wip: cleaning up the code, moving graph ts, refining the stylistic choices
