+++ src/elm/Util.elm
@@ -11,6 +11,7 @@ module Util exposing
     , base64Decode
     , boolToYesNo
     , buildRefURL
+    , formatDuration
     , checkScheduleAllowlist
     , dispatch
     , extractFocusIdFromRange
@@ -23,6 +24,7 @@ module Util exposing
     , getNameFromRef
     , humanReadableDateTimeFormatter
     , humanReadableWithDefault
+    ,formatRunTimeWithDefault
     , isLoading
     , isSuccess
     , largeLoader
@@ -159,6 +161,20 @@ noSomeSecondsAgo _ =
     "just now"
 
 
+
+{-| formatDuration : calculates build runtime using a duration
+-}
+formatDuration : Int -> String
+formatDuration duration =
+    let
+        minutes =
+            runTimeMinutes <| duration  
+
+        seconds =
+            runTimeSeconds <| duration 
+    in
+    String.join ":" [ minutes, seconds ]
+
 {-| formatRunTime : calculates build runtime using current application time and build times
 -}
 formatRunTime : Posix -> Int -> Int -> String
@@ -174,6 +190,26 @@ formatRunTime now started finished =
             runTimeSeconds runtime
     in
     String.join ":" [ minutes, seconds ]
+
+
+{-| formatRunTimeWithDefault : calculates build runtime using current application time and build times, with a default for non-started times
+-}
+formatRunTimeWithDefault : Posix -> Int -> Int -> String
+formatRunTimeWithDefault now started finished =
+    if started == 0 then
+        "--:--"
+    else
+        let
+            runtime =
+                buildRunTime now started finished
+
+            minutes =
+                runTimeMinutes runtime
+
+            seconds =
+                runTimeSeconds runtime
+        in
+        String.join ":" [ minutes, seconds ]
 
 
 {-| buildRunTime : calculates build runtime using current application time and build times, returned in seconds
