+++ src/scss/_graph.scss
@@ -193,19 +193,26 @@
 }
 
 .d3-build-graph-node-step-a {
-  outline: none;
-
-  &.-hover {
-    // todo: finalize step link styling and view data (ruleset info?)
-    //  its difficult to see step information
-    outline: 1px solid var(--color-secondary);
-    outline-style: dashed;
+  &:focus,&:active {
+    outline:none;
   }
 
+  &.-hover {
+    text {
+      fill: var(--color-primary);
+    }
+  }
+  
   // used?
   &.-filtered {
     outline: 1px solid var(--color-primary);
     outline-style: dashed;
+  }
+}
+
+.d3-build-graph-node-step-a-underline {
+  &.-hover {
+    fill: var(--color-primary);
   }
 }
