+++ src/static/index.ts
@@ -13,7 +13,7 @@ import { Graphviz } from '@hpcc-js/wasm';
 import { Elm } from '../elm/Main.elm';
 import '../scss/style.scss';
 import { App, Config, Flags, Theme } from './index.d';
-
+import * as Graph from  './graph';
 
 // Vela consts
 const feedbackURL: string =
@@ -133,433 +133,25 @@ function envOrNull(env: string, subst: string): string | null {
   return subst;
 }
 
-// track current rendering build
-var currentBuild = -1;
-var isRefreshDraw = false;
-var contentFilter = "";
+// track rendering options globally to help determine draw logic
+var opts = {
+  currentBuild: -1,
+  isRefreshDraw: false,
+  contentFilter: "",
+};
 
 app.ports.renderBuildGraph.subscribe(function (graphData) {
   const dot = graphData.dot;
+
   const graphviz = Graphviz.load().then(res => {
     var content = res.layout(dot, 'svg', 'dot');
-    isRefreshDraw = (currentBuild === graphData.build_id);
-    currentBuild = graphData.build_id;
-    contentFilter = graphData.filter;
-    drawGraph(content);
+
+    // construct graph building options
+    opts.isRefreshDraw = (opts.currentBuild === graphData.build_id);
+    opts.currentBuild = graphData.build_id;
+    opts.contentFilter = graphData.filter;
+    opts.onGraphInteraction = app.ports.onGraphInteraction;
+
+    Graph.drawGraph(opts, content);
   });
 });
-
-function drawGraph(content) {
-  // force d3 exports to resolve, required or we receive errors when running within a container
-  // this is why we love javascript
-  var _ = d3;
-
-  // define DOM selectors for DOT-generated elements
-  var graphSelectors = {
-    root: '.elm-build-graph-root',
-    node: '.elm-build-graph-node',
-    edge: '.elm-build-graph-edge',
-  };
-
-  var buildGraphElement = drawBaseGraph(graphSelectors.root, content);
-
-  // check that a valid graph was rendered
-  if (buildGraphElement.node() == null) {
-    console.log('unable to continue drawing graph, root element is invalid');
-    console.log(buildGraphElement);
-    return;
-  }
-
-  drawViewbox(buildGraphElement);
-
-  // apply onclick to base node links prior to adding/removing elements
-  applyNodesOnClick(buildGraphElement, graphSelectors.node);
-
-  var edges = drawEdges(buildGraphElement, graphSelectors.edge);
-
-  drawNodes(buildGraphElement, graphSelectors.node, edges);
-}
-
-function drawBaseGraph(selector, content) {
-  // grab the build graph root element
-  var buildGraphElement = d3.select(selector);
-  var graphBBox = buildGraphElement.node().getBBox();
-
-  let height = 800;
-  let width = 1500;
-
-  var zoom = d3.zoom()
-    .scaleExtent([0.1, Infinity])
-    // .translateExtent([0,0], [])
-    .on('zoom', handleZoom);
-
-  // define d3 zoom function
-  function handleZoom(event) {
-    if (isNaN(event.transform.k)) {
-      event.transform.k = 1;
-    }
-    if (isNaN(event.transform.x)) {
-      event.transform.x = 0;
-    }
-    if (isNaN(event.transform.y)) {
-      event.transform.y = 0;
-    }
-
-    var zoomG = d3.select(selector + ' g');
-    zoomG
-      .attr('transform', event.transform);
-  }
-
-  function resetZoomAndCenter(zoom) {
-    console.log("resetting zoom and centering");
-    var zoomG = d3.select(selector);
-    zoomG
-    .call(zoom.scaleTo, 1);
-    var zoomGg = d3.select(selector);
-    var w = zoomGg.node().getBBox().width;
-    var h = zoomGg.node().getBBox().height;
-    zoomGg
-    .transition() // required to 'chain' these two instant animations together
-    .duration(0)
-    .call(zoom.translateTo, w * 0.5, h * 0.5)
-  }
-
-  // enable d3 zoom and pan functionality
-  buildGraphElement.call(zoom);
-
-  var actionResetPan = d3.select('#action-center');
-
-  actionResetPan.on('click', function (e) {
-    e.preventDefault();
-    console.log('center: onclick');
-    resetZoomAndCenter(zoom);
-  });
-
-  // apply mousedown zoom effects
-  var g = d3.select('g.node_mousedown');
-  if (g.empty()) {
-    g = buildGraphElement
-      .append('g')
-      .classed('node_mousedown', true)
-      .attr('id', 'zoom');
-  }
-
-  // this centers the graph in the viewbox, or something like that
-  buildGraphElement = g;
-
-  // draw content into html
-  buildGraphElement.html(content);
-
-  if (!isRefreshDraw) {
-    console.log("resetting the graph because build number changed");
-    resetZoomAndCenter(zoom);
-  }
-  
-  return buildGraphElement;
-}
-
-function drawViewbox(buildGraphElement) {
-  var graphBBox = buildGraphElement.node().getBBox();
-
-  // apply viewbox properties to the root element's parent
-  // provide x padding for the legend
-  const VIEWBOX_PADDING = { x1: 0, x2: 500, y1: 0, y2: 100 };
-
-  var graphParent = d3.select(buildGraphElement.node().parentNode);
-  graphParent.attr(
-    'viewBox',
-    '' +
-    (graphBBox.x - VIEWBOX_PADDING.x1) +
-    ' ' +
-    (graphBBox.y - VIEWBOX_PADDING.y1) +
-    ' ' +
-    (graphBBox.width + VIEWBOX_PADDING.x2) +
-    ' ' +
-    (graphBBox.height + VIEWBOX_PADDING.y2),
-  );
-}
-
-function drawNodes(buildGraphElement, selector, edges) {
-  buildGraphElement.selectAll(selector).filter(function () {
-    let stageNode = d3.select(this);
-    var nodeBBox = stageNode.node().getBBox();
-
-    // apply an outline using rect, since nodes are rect and this will allow for animation
-    var outline = stageNode.append('rect');
-    outline
-      .attr('x', nodeBBox.x)
-      .attr('y', nodeBBox.y)
-      .attr('width', nodeBBox.width)
-      .attr('height', nodeBBox.height);
-
-    var stageInfo = stageNode.attr('id').replace('#', '').split(',');
-
-    // todo: safety-check
-    var stageName = stageInfo[1];
-    var stageStatus = stageInfo[2];
-
-    // restore base class and build modifiers
-    outline.attr('class', 'd3-build-graph-node-outline-rect');
-
-    var restoreNodeClass = o => {
-      o.classed('-pending', true);
-    };
-
-    if (stageStatus === 'failure') {
-      restoreNodeClass = o => {
-        o.classed('-failure', true);
-      };
-    }
-
-    if (stageStatus === 'success') {
-      restoreNodeClass = o => {
-        o.classed('-success', true);
-      };
-    }
-
-    if (stageStatus === 'running') {
-      restoreNodeClass = o => {
-        o.classed('-running', true);
-      };
-    }
-
-    if (stageStatus === 'killed') {
-      restoreNodeClass = o => {
-        o.classed('-killed', true);
-      };
-    }
-
-    if (contentFilter.length > 2 && stageName.includes(contentFilter)) {
-      var size = 15;
-      size = 10;
-      var outline2 = stageNode.append('rect');
-      outline2.attr('class', 'd3-build-graph-node-outline-rect');
-
-      outline2
-        .attr('x', nodeBBox.x - (size * 0.5))
-        .attr('y', nodeBBox.y - (size * 0.5))
-        .attr('width', nodeBBox.width + size)
-        .attr('height', nodeBBox.height + size)
-        .classed('-filtered', true);
-    }
-
-    restoreNodeClass(outline);
-
-    stageNode.on('mouseover', e => {
-      outline.classed('-hover', true);
-
-      // take this stage
-      // filter out all the edges that arent source/dest of each edge
-      edges.filter(function (edgeInfo) {
-        if (
-          // todo: safety-check
-          stageInfo[0] === edgeInfo.source ||
-          stageInfo[0] === edgeInfo.destination
-        ) {
-          edgeInfo.target.classed('-hover', true);
-        }
-      });
-    });
-
-    stageNode.on('mouseout', e => {
-      // remove node outline styling
-      outline.classed('-hover', false);
-
-      // restore node styling
-      restoreNodeClass(outline);
-
-      // restore edge styling
-      edges.filter(function (edgeInfo) {
-        // modify styling only on related edges
-        if (
-          // todo: safety-check
-          stageInfo[0] === edgeInfo.source ||
-          stageInfo[0] === edgeInfo.destination
-        ) {
-          var status = edgeInfo.status;
-
-          edgeInfo.target.classed('-hover', false);
-
-          var restoreEdgeClass = o => {
-            o.classed('-pending', true);
-          };
-
-          if (status === 'running') {
-            restoreEdgeClass = o => {
-              o.classed('-running', true);
-            };
-          }
-
-          if (status === 'success') {
-            restoreEdgeClass = o => {
-              o.classed('-success', true);
-            };
-          }
-
-          if (status === 'failure') {
-            restoreEdgeClass = o => {
-              o.classed('-failure', true);
-            };
-          }
-
-          restoreEdgeClass(edgeInfo.target);
-        }
-      });
-    });
-
-    stageNode.selectAll('a').filter(function () {
-      var step = d3.select(this);
-      if (step.attr('xlink:href').includes('#step')) {
-        // restore base class and build modifiers
-        step.attr('class', 'd3-build-graph-node-step-a');
-
-        step.on('mouseover', e => {
-          step.classed('-hover', true);
-        });
-        step.on('mouseout', e => {
-          step.classed('-hover', false);
-        });
-      }
-    });
-
-    stageNode.selectAll('#a_node-cell').filter(function () {
-      var cell = d3.select(this);
-      var cellNode = cell.select('text').node();
-      if (cellNode) {
-        let parent = d3.select(cellNode.parentNode);
-        let nodeBox = cellNode.getBBox();
-
-        // extract href to dispatch to Elm
-        var href = parent.attr('xlink:href');
-
-        // remove actual href attribute
-        parent.attr('xlink:href', null);
-
-        var stepInfo = parent.attr('xlink:title').split(',');
-
-        // todo: safety check
-        var status = stepInfo[2];
-
-        // todo: static/*.png seems like a bad way to do icon images
-        parent
-          .append('image')
-          .attr('xlink:href', '/images/vela_' + status + '.png')
-          .attr('x', nodeBox.x - 6)
-          .attr('y', nodeBox.y)
-          .attr('width', 16)
-          .attr('height', 16);
-
-        parent.on('click', function (e) {
-          e.preventDefault();
-
-          // prevents multiple link events getting fired from a single click
-          e.stopImmediatePropagation();
-
-          setTimeout(
-            () =>
-              app.ports.onGraphInteraction.send({
-                event_type: 'href',
-                href: href,
-                step_id: '',
-              }),
-            0,
-          );
-        });
-      }
-    });
-
-    return ''; // used by filter (?)
-  });
-}
-
-// applyNodesOnClick takes root graph element, selects node links and applies onclick functionality
-function applyNodesOnClick(buildGraphElement, selector) {
-  // process and return all 'linked' stage nodes
-  return buildGraphElement.selectAll(selector + ' a').filter(function () {
-    // todo: figure out .each
-    // add onclick to nodes with valid href attributes
-    var href = d3.select(this).attr('xlink:href');
-    if (href !== null) {
-      d3.select(this).on('click', function (e) {
-        e.preventDefault();
-        var nodeA = d3.select(this);
-        // extract identifier from href
-        // todo: make this use title
-        var data = nodeA.attr('xlink:href');
-        nodeA.attr('xlink:href', null);
-        let id = data.replace('#', '');
-        setTimeout(
-          () =>
-            app.ports.onGraphInteraction.send({
-              event_type: 'node_click',
-              node_id: id,
-            }),
-          0,
-        );
-      });
-    }
-    return '';
-  });
-}
-
-function drawEdges(buildGraphElement, selector) {
-  // collect edge information to use in other d3 interactivity
-  var edges = [];
-
-  buildGraphElement.selectAll(selector).filter(function () {
-    let a = d3.select(this);
-    var edgeInfo = a.attr('id').replace('#', '').split(',');
-
-    // todo: safety-check
-    var status = edgeInfo[2];
-
-    var p = a.select('path');
-
-    edges.push({
-      target: p,
-      source: edgeInfo[0], // todo: safety-check
-      destination: edgeInfo[1], // todo: safety-check
-      status: status,
-    });
-
-    // restore base class and build modifiers
-    p.attr('class', 'd3-build-graph-edge-path');
-
-    var restoreEdgeClass = o => {
-      o.classed('-pending', true);
-    };
-
-    if (status === 'running') {
-      restoreEdgeClass = o => {
-        o.classed('-running', true);
-      };
-    }
-
-    if (status === 'success') {
-      restoreEdgeClass = o => {
-        o.classed('-success', true);
-      };
-    }
-
-    if (status === 'failure') {
-      restoreEdgeClass = o => {
-        o.classed('-failure', true);
-      };
-    }
-
-    restoreEdgeClass(p);
-
-    a.on('mouseover', e => {
-      p.classed('-hover', true);
-    });
-
-    a.on('mouseout', e => {
-      restoreEdgeClass(p);
-      p.classed('-hover', false);
-    });
-
-    return ''; // used by filter (?)
-  });
-
-  return edges;
-}
